Upload media assets added to previously synced observations.
We track the set of ML IDs associated with an observation in the description field. This is somewhat fragile; it would be better to use the "orginal filename" field of the uploaded photo or sound, but the iNaturalist API doesn't expose this. I have raised this for discussion on the iNaturalist forum (https://forum.inaturalist.org/t/api-return-original-filename-when-fetching-photos-or-sounds/69322/5) and filed a PR to implement it (https://github.com/inaturalist/iNaturalistAPI/pull/534), but this has not been reviewed or merged. This CL is a workaround.

This CL makes a variety of other code cleanups and refactorings.